fix: 클라이언트 데이터 수집 문제 해결
- API 엔드포인트 경로 수정 (/collect → /api/analytics/collect)
- 백엔드 주소 업데이트
- CORS 설정 추가 (whwnddml.github.io, brandiup.com)
- PostgreSQL 시간 필드 타입을 BIGINT로 변경

+++ backend/database/init/01-create-tables.sql
@@ -37,11 +37,11 @@ CREATE TABLE IF NOT EXISTS area_engagements (
     area_id VARCHAR(50) NOT NULL,
     area_name VARCHAR(100) NOT NULL,
     area_type VARCHAR(50),
-    time_spent INTEGER,
+    time_spent BIGINT,
     interaction_count INTEGER DEFAULT 0,
     first_engagement TIMESTAMP WITH TIME ZONE,
     last_engagement TIMESTAMP WITH TIME ZONE,
-    visible_time INTEGER,
+    visible_time BIGINT,
     viewport_percent NUMERIC(5,2),
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
